requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
opencv-python==4.10.0.84
numpy==2.1.1
Pillow==10.4.0
python-multipart==0.0.9
